+++ src/pages/_components/Image.svelte
@@ -1,26 +1,60 @@
 <script>
-    export let src;
-    export let alt;
-    export let cls;
-    export let full;
+    export let file;
+    export let post_view = false;
     import { onDestroy } from "svelte";
 
-    let hoverEl = document.createElement("img");
+    // Initiate file
+    let hoverEl,
+        thumbnail,
+        alt,
+        view,
+        expanded = false;
+    let isVideo = file.mime_type === "video/webm";
+    if (isVideo) {
+        hoverEl = document.createElement("video");
+        hoverEl.loop = true;
+        thumbnail =
+            "/thumbnails/" +
+            file.file_path
+                .split(".")
+                .slice(0, -1)
+                .join(".") +
+            ".jpg";
+    } else {
+        hoverEl = document.createElement("img");
+        thumbnail = "/thumbnails/" + file.file_path;
+    }
+
+    hoverEl.src = "/uploads/" + file.file_path;
     hoverEl.className = "imageHover";
-    let id = Math.floor(Math.random() * 1000000);
+    view = thumbnail;
+    alt = file.file_name;
     $: {
-        hoverEl.id = "img_" + id;
-        hoverEl.src = full;
+        console.log("post_view", post_view);
+        if (expanded) {
+            view = "/uploads/" + file.file_path;
+        } else {
+            view = thumbnail;
+        }
     }
 
     const imgHover = el => {
-        console.log(hoverEl.src);
         document.body.appendChild(hoverEl);
+        if (isVideo) {
+            hoverEl.play();
+        }
     };
 
     const imgCleanup = el => {
         hoverEl.remove();
     };
+
+    const imgExpand = el => {
+        console.log("CLICKED", el);
+        hoverEl.remove();
+        expanded = !expanded;
+    };
+
     onDestroy(() => {
         imgCleanup();
     });
@@ -43,11 +77,55 @@
     img {
         object-fit: contain;
     }
+    .playback {
+        position: relative;
+        .contract {
+            position: absolute;
+            top: 5px;
+            left: 100%;
+        }
+    }
 </style>
 
-<img
-    class={cls}
-    {src}
-    alt={alt ?? 'This is an image'}
-    on:mouseenter={imgHover}
-    on:mouseleave={imgCleanup} />
+{#if !post_view}
+    <img
+        id={'img_' + file.id}
+        class={''}
+        src={thumbnail}
+        alt={alt ?? 'This is an image'}
+        on:mouseenter={imgHover}
+        on:mouseleave={imgCleanup} />
+{:else}
+    {#if !expanded}
+        <img
+            id={'img_' + file.id}
+            class={'single_post__image'}
+            src={view}
+            alt={alt ?? 'This is an image'}
+            on:mouseenter={imgHover}
+            on:click={imgExpand}
+            on:mouseleave={imgCleanup} />
+    {:else}
+        <!-- asdfasdfasdfasdf -->
+        {#if isVideo}
+            <div class="playback single_post__image expanded" id={'img_' + file.id}>
+                <!-- svelte-ignore a11y-media-has-caption -->
+                <video src={view} controls autoplay />
+                <!-- svelte-ignore a11y-missing-attribute -->
+                <a class="icon is-small contract" on:click={imgExpand}>
+                    <i class="fas fa-times fa-inverse" />
+                </a>
+            </div>
+        {:else} 
+            <img
+            id={'img_' + file.id}
+            class={'single_post__image expanded'}
+            src={view}
+            alt={alt ?? 'This is an image'}
+            style="min-height: 600px"
+            on:mouseenter={imgHover}
+            on:click={imgExpand}
+            on:mouseleave={imgCleanup} />
+        {/if}
+    {/if}
+{/if}